+++ src/exceptions.ts
@@ -1,0 +1,1 @@
+export const INVALID_DATACHANNEL_CALLBACK = 'Must provide a valid callback';